Fix bug in audiobuffersource-duration-loop test
The test was producing a failure that the array lengths didn't match.
(Not sure why the test was marked as passing even though it failed.)
The problem was that we actually had audioBuffer.getChannelData
instead of audioBuffer.getChannelData(0) to get the data from the
first channel.

Correct the mistake and also run clang-format over the code to clean
it up.

Bug: 1217791
Change-Id: I510838337e2948b89fb0622b63dd81e022d103d0