[CLS] Allow for shifts in root-relative coordinates for scroll anchoring
Currently, layout adjustments plus scroll anchoring may result in all
three of the following happening in the same rendering update:

* Layout changes that moves a tracked element relative to its
layout shift root (e.g. via content-visibility changing skip state, but
it could also be any layout change offscreen)
* Natural scroll via a user gesture that visually moves the element
visually on-screen
* Scroll anchoring to counter the layout change, but not the natural
scroll

We currently do not handle that case. We only handle the case of
a layout change that is directly counter-balanced by a scroll, or
a pure scroll.

This CL serves the above use case by plumbing the value of the scroll
anchor scroll to LayoutShiftTracker, and early-outing if the root-
relative position of the before and after rects are equal, once
scroll anchoring is taken into account.

Bug: 1215368

Change-Id: I8a5240603cebf3a78dfc4fa82b4839cf1af68800